move toDecimalString to bigint util and drop duplicate tests
