--- app/util/formattedDeeplinkParsedValue.test.ts
@@ -1,16 +1,0 @@
-import formattedDeeplinkParsedValue from './formattedDeeplinkParsedValue';
-
-describe('formattedDeeplinkParsedValue', () => {
-  it.each([
-    ['1e+21', '1000000000000000000000'],
-    ['1.234e+17', '123400000000000000'],
-    ['1.4e+18', '1400000000000000000'],
-    ['1.6e+21', '1600000000000000000000'],
-    ['4.2e+17', '420000000000000000'],
-    ['1000000', '1000000'],
-    // Verifies no float precision loss — Number() would corrupt this to 10000000000000000000000
-    ['9999999999999999999999', '9999999999999999999999'],
-  ])('converts %s to %s', (input, expected) => {
-    expect(formattedDeeplinkParsedValue(input)).toEqual(expected);
-  });
-});
